protocol: add true and false response types
Added new constants TRUE and FALSE to the ResponseType enum. The
original approach considered was to introduce a new 'code' field which
would store these response codes. However, since the response type
already has so much empty space, it does not make much sense to
introduce a new field.

Therefore, the ResponseType byte will contain TRUE and FALSE directly
and can be extended in the future to support other possible responses.

This commit introduces the required response types necessary for the
implementation of #5 and defines the approach that will be needed to
address #2